refactor OpsLifecycleAdapter
