+++ pkg/controllers/operationjob/restart/restart_lifecycle_adapter.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package utils
+package restart
 
 import (
 	"sigs.k8s.io/controller-runtime/pkg/client"
@@ -23,7 +23,7 @@ import (
 )
 
 var (
-	RestartOpsLifecycleAdapter = &OperationJobRestartOpsLifecycleAdapter{}
+	OpsLifecycleAdapter = &OperationJobRestartOpsLifecycleAdapter{}
 )
 
 // OperationJobRestartOpsLifecycleAdapter tells PodOpsLifecycle the container restart ops info
